AI Bot/requirements.txt
ccxt
pandas
ta
python-telegram-bot
python-dotenv
colorama
openpyxl
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0
numpy
customtkinter
CTkDateEntry>=0.1.2
pyinstaller
